Add the final bulk changes
API works with bulk upload and response CSV gives correct data
