+++ src/main/services/user-auth.ts
@@ -4,7 +4,6 @@ import { HTTPError } from '../HttpError';
 
 export function getUserAccessToken(req: Request): string {
   const accessToken = req.oidc?.accessToken?.access_token;
-
   if (!accessToken) {
     throw new HTTPError('Forbidden', 403);
   }
